third_party/WebKit/LayoutTests/inspector-protocol/css/cssom-matching-rules-multiple-expected.txt

The test verifies CSS.getMatchedStylesForNode when used concurrently with multiple CSSOM modifications.
Running test: testModifyRule
--------------
Original rule:
*#test* {    regular
    color: red;
}
*#test* {    regular
    color: green;
}
*#test* {    regular
    color: blue;
}
*#test* {    regular
    width: 10%;
}
*#test* {    regular
    width: 20%;
}
*#test* {    regular
    width: 30%;
}
*#test* {    regular
    width: 40%;
}
*#test* {    regular
    width: 50%;
}
*#test* {    regular
    width: 60%;
}
Mutating 3rd:
---------------
*#test* {    regular
    color: red;
}
*#test* {    regular
    color: green;
}
*#test* {    regular
    color: blue;
}
*#test* {    regular    readonly
    color: red;
}
*#test* {    regular
    width: 20%;
}
*#test* {    regular
    width: 30%;
}
*#test* {    regular
    width: 40%;
}
*#test* {    regular
    width: 50%;
}
*#test* {    regular
    width: 60%;
}
Mutating 4th:
--------------
*#test* {    regular
    color: red;
}
*#test* {    regular
    color: green;
}
*#test* {    regular
    color: blue;
}
*#test* {    regular    readonly
    color: red;
}
*#test* {    regular    readonly
    color: green;
}
*#test* {    regular
    width: 30%;
}
*#test* {    regular
    width: 40%;
}
*#test* {    regular
    width: 50%;
}
*#test* {    regular
    width: 60%;
}
Mutating 5th:
--------------
*#test* {    regular
    color: red;
}
*#test* {    regular
    color: green;
}
*#test* {    regular
    color: blue;
}
*#test* {    regular    readonly
    color: red;
}
*#test* {    regular    readonly
    color: green;
}
*#test* {    regular    readonly
    color: blue;
}
*#test* {    regular
    width: 40%;
}
*#test* {    regular
    width: 50%;
}
*#test* {    regular
    width: 60%;
}
Delete first 3:
---------------
*#test* {    regular
    color: red;
}
*#test* {    regular
    color: green;
}
*#test* {    regular
    color: blue;
}
*#test* {    regular
    width: 40%;
}
*#test* {    regular
    width: 50%;
}
*#test* {    regular
    width: 60%;
}
